chore: create a use-cases package to provide new errors to transfer's repository
