+++ pkg/domain/transfers/usecases/update_balance_test.go
@@ -1,4 +1,4 @@
-package transfers
+package usecases
 
 import (
 	"context"
